requirements.txt
streamlit
duckdb
getdaft
requests
google-generativeai
pandas
python-dotenv